backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-core==1.34.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.1
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.3
contourpy==1.3.0
cryptography==45.0.3
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.3
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dlib==20.0.0
exceptiongroup==1.3.0
executing==2.2.1
face-recognition==1.3.0
face_recognition_models==0.3.0
fastjsonschema==2.21.2
filelock==3.18.0
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel==6.31.0
ipython==8.18.1
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.7
lark==1.3.1
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.2.1
mistune==3.2.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
narwhals==2.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.5.1
notebook_shim==0.2.4
numpy==2.0.2
openai==0.28.0
opencv-python==4.11.0.86
oscrypto==1.3.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pillow==11.2.1
platformdirs==4.4.0
premailer==3.10.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.1
protobuf==6.33.0
psutil==7.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==5.7.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==4.0.0
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.2
pyzmq==27.1.0
razorpay==2.0.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
safetensors==0.5.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.50.0
svglib==1.5.1
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.3.0
torch==2.8.0
torchaudio==2.7.0+cpu
torchvision==0.22.0+cpu
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.38.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritools==5.0.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
xhtml2pdf==0.2.17
xlsxwriter==3.2.5
yagmail==0.15.293
yarl==1.20.0
zipp==3.23.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
razorpay==1.4.1
python-dotenv==1.0.0
Pillow==10.0.0